Switching tabs in the Chat shows metamask popup
